move execution plan viewer code to @finos/legend-query-builder
