+++ .changeset/lazy-apricots-yawn.md
@@ -1,0 +1,7 @@
+---
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-application-studio': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-application': patch
+'@finos/legend-art': patch
+---
