+++ .changeset/many-moose-unite.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-application': major
+---
+
+**BREAKING CHANGE:** Moved execution plan viewer logic to `@finos/legend-query-builder`
