+++ packages/legend-application/package.json
@@ -59,11 +59,9 @@
     "react-dnd": "16.0.1",
     "react-dnd-html5-backend": "16.0.1",
     "react-dom": "18.2.0",
-    "react-draggable": "4.4.5",
     "react-router": "5.3.4",
     "react-router-dom": "5.3.4",
     "serializr": "3.0.2",
-    "sql-formatter": "12.2.0",
     "xterm": "5.1.0",
     "xterm-addon-fit": "0.7.0",
     "xterm-addon-search": "0.11.0",
