+++ packages/legend-application/src/index.ts
@@ -61,10 +61,6 @@ export { DISPLAY_ANSI_ESCAPE } from './stores/terminal/Terminal.js';
 
 // ------------------------------------------- TO BE MOVED -------------------------------------------
 
-// TODO: move to `@finos/legend-query-builder`
-export * from './components/execution-plan-viewer/ExecutionPlanViewer.js';
-export * from './stores/ExecutionPlanState.js';
-
 // TODO: move to `@finos/legend-graph-editor`
 export {
   PURE_GRAMMAR_TOKEN,
