+++ packages/legend-art/package.json
@@ -64,6 +64,7 @@
     "react-dnd": "16.0.1",
     "react-dnd-html5-backend": "16.0.1",
     "react-dom": "18.2.0",
+    "react-draggable": "4.4.5",
     "react-icons": "4.8.0",
     "react-markdown": "8.0.6",
     "react-reflex": "4.0.12",
