+++ packages/legend-query-builder/style/_execution-plan-viewer.scss
@@ -1,0 +1,327 @@
+/**
+ * Copyright (c) 2020-present, Goldman Sachs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@use 'mixins' as *;
+
+.execution-plan-viewer {
+  height: 100%;
+  width: 100%;
+
+  &__content {
+    height: 100%;
+    width: 100%;
+  }
+
+  &__debug-panel {
+    background: var(--color-dark-grey-50);
+
+    .panel__header {
+      background: var(--color-dark-grey-250);
+    }
+
+    .panel__content {
+      position: relative;
+      background: var(--color-dark-grey-50);
+    }
+  }
+
+  &__close-btn {
+    border-color: var(--color-blue-200);
+    background: var(--color-blue-300);
+  }
+
+  .panel {
+    height: 100%;
+
+    &__header {
+      color: var(--color-light-grey-400);
+      font-weight: bold;
+      background: var(--color-dark-grey-50);
+
+      &__title {
+        max-width: calc(100% - 11rem);
+      }
+    }
+
+    &__main-header {
+      color: var(--color-light-grey-400);
+      font-weight: bold;
+      background: var(--color-dark-grey-250);
+    }
+
+    &__content {
+      position: relative;
+      height: calc(100% - 3rem);
+      width: 100%;
+      background: var(--color-dark-grey-50);
+      overflow-y: hidden;
+
+      & ::-webkit-scrollbar-thumb {
+        background: var(--color-dark-grey-400);
+      }
+    }
+
+    &__content > div:first-child {
+      margin-top: 0;
+    }
+
+    &__content__lists {
+      height: 100%;
+      width: 100%;
+      padding: 1rem;
+      overflow: overlay;
+    }
+  }
+
+  &__tabs__header {
+    padding-left: 0;
+  }
+
+  &__tabs {
+    height: 2.8rem;
+    z-index: 1;
+    display: flex;
+    overflow-x: overlay;
+    overflow-y: hidden;
+  }
+
+  &__tab {
+    height: 100%;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0 1rem;
+    border-right: 0.1rem solid var(--color-dark-grey-200);
+    white-space: nowrap;
+    cursor: pointer;
+  }
+
+  &__tab--active {
+    color: var(--color-dark-grey-50);
+    position: relative;
+  }
+
+  &__tab--active::after {
+    content: '';
+    height: 0.4rem;
+    width: 100%;
+    position: absolute;
+    bottom: 0;
+    background: var(--color-yellow-200);
+  }
+
+  &__panel {
+    height: 100%;
+    width: 100%;
+
+    &__content {
+      overflow-y: hidden;
+      position: relative;
+      background: var(--color-dark-grey-50);
+      height: calc(100% - 3.4rem);
+    }
+
+    &__header {
+      background: var(--color-dark-grey-100);
+      z-index: 1;
+      height: 3.4rem;
+      min-height: 3.4rem;
+      padding: 0;
+    }
+
+    &__header__tabs {
+      display: flex;
+      height: 100%;
+      overflow-x: overlay;
+    }
+
+    &__header__tab {
+      @include flexVCenter;
+
+      padding: 0 1rem;
+      cursor: pointer;
+      color: var(--color-light-grey-400);
+      background: var(--color-dark-grey-80);
+      border-right: 0.1rem solid var(--color-dark-grey-80);
+    }
+
+    &__header__tab--active {
+      color: var(--color-light-grey-100);
+      background: var(--color-dark-grey-50);
+    }
+
+    &__header__tab__content {
+      @include flexVCenter;
+
+      height: 100%;
+      width: 100%;
+    }
+
+    &__header__tab__icon {
+      margin-right: 0.5rem;
+    }
+
+    &__header__tab__label {
+      color: inherit;
+      padding: 0 0.5rem;
+      white-space: nowrap;
+    }
+
+    &__view-mode {
+      &__type {
+        @include flexCenter;
+
+        width: 15rem;
+        background: var(--color-dark-grey-100);
+        border-right: 0.1rem solid var(--color-dark-grey-50);
+        border-left: 0.1rem solid var(--color-dark-grey-50);
+        color: var(--color-light-grey-300);
+        cursor: pointer;
+      }
+
+      &__type:hover {
+        background: var(--color-dark-grey-200);
+      }
+
+      &__type__label {
+        @include flexVCenter;
+
+        height: 3.4rem;
+        padding: 0 0.5rem;
+        border-top: 0.3rem solid var(--color-yellow-200);
+      }
+
+      &__options {
+        width: 15rem;
+      }
+
+      &__options--with-group {
+        border: none;
+        padding: 0;
+      }
+
+      &__option__group__name {
+        @include flexCenter;
+
+        flex-direction: column;
+        cursor: default;
+        writing-mode: vertical-lr;
+        transform: rotate(180deg);
+        font-size: 1.1rem;
+        width: 2.2rem;
+        padding: 0.5rem 0;
+      }
+
+      &__option__group {
+        display: flex;
+        flex-direction: row;
+
+        &--native {
+          background: var(--color-blue-100);
+          border: 0.1rem solid var(--color-blue-100);
+        }
+
+        &--native &__name {
+          background: var(--color-blue-100);
+          color: var(--color-light-grey-50);
+        }
+
+        &--generation {
+          background: var(--color-pink-300);
+          border: 0.1rem solid var(--color-pink-300);
+        }
+
+        &--generation &__name {
+          background: var(--color-pink-300);
+          color: var(--color-light-grey-50);
+        }
+      }
+
+      &__option__group__separator {
+        height: 0.2rem;
+        width: 100%;
+      }
+
+      &__option__group__options {
+        width: 100%;
+        background: var(--color-dark-grey-100);
+      }
+
+      &__option {
+        @include flexCenter;
+
+        height: 2rem;
+      }
+    }
+  }
+
+  &__explorer {
+    width: 100%;
+
+    &__content__container {
+      height: 100%;
+      background: var(--color-dark-grey-100);
+    }
+  }
+
+  &__explorer-tree__node__container:hover {
+    background: var(--color-dark-blue-shade-100);
+  }
+
+  &__explorer-tree__node__container--selected,
+  &__explorer-tree__node__container--selected:hover {
+    background: var(--color-light-blue-450);
+  }
+
+  &__explorer-tree__node__icon {
+    width: 4rem;
+    min-width: 4rem;
+  }
+
+  &__explorer-tree__node__icon__expand,
+  &__explorer-tree__node__icon__type {
+    width: 2rem;
+
+    @include flexHCenter;
+  }
+
+  &__explorer-tree__node__icon__expand svg {
+    font-size: 1rem;
+  }
+
+  &__explorer-tree__node__label {
+    color: inherit;
+  }
+}
+
+.execution-plan-viewer,
+.execution-node-viewer {
+  &__unsupported-view {
+    @include flexCenter;
+
+    flex-direction: column;
+
+    &__summary {
+      font-weight: 500;
+      margin-bottom: 1rem;
+    }
+
+    &__to-text-mode__btn {
+      height: 2.8rem;
+      width: 15rem;
+    }
+  }
+}
